backend/requirements/base.txt
channels==4.0.0
channels_redis==4.1.0
daphne==4.0.0
Django==4.0
django-cors-headers==4.3.0
django-filter==23.3
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Pillow==10.1.0
psycopg==3.1.8
redis==5.0.1
django-redis==5.4.0
requests==2.31.0
psycopg2-binary
drf-spectacular